requirements.txt
Django==4.1.3
Pillow==9.3.0
django-import-export==3.0.1
djangorestframework==3.14.0
django-filter==22.1
django-cors-headers==3.13.0